feat and refactor: Implementações calculos para retorno do total e cashback com base no total de produtos, ajustes nas classes para efetivação e novo request de retorno, organização dos imports não utilizados

+++ cashsplash/src/main/java/com/example/cashsplash/dtos/campanha/CampaignRequestDTO.java
@@ -12,5 +12,5 @@ import java.math.BigDecimal;
 public class CampaignRequestDTO {
     String name;
     String description;
-    Double offValue;
+    BigDecimal offValue;
 }
